requirements/tests.txt
attrs==19.1.0
flake8==3.7.7
pytest==4.5.0
